+++ DAO/EquipeDAO.php
@@ -113,5 +113,5 @@ class EquipeDAO
         return $req;
     }
 
-
+    
 }
